+++ app/routes/test/test-ctrl.js
@@ -1,7 +1,7 @@
 'use strict';
 
 angular.module('cmRoutes').controller('TestCtrl',
-    function ($scope, $rootScope, $q, $timeout, $interval) {
+    function ($scope, $rootScope, $q, $timeout, $interval, cmModal) {
 
         $rootScope.$broadcast('device:goesToBackground')