+++ core/comps/contacts/srvc-contact-model.js
@@ -2,15 +2,12 @@
 
 angular.module('cmContacts')
 .factory('cmContactModel', [
-    'cmContactsAdapter',
-    'cmIdentityFactory',
-    'cmObject',
-    'cmStateManagement',
-    'cmUtil',
-    'cmLogger',
-    'cmSecurityAspectsContact',
+    'cmContactsAdapter', 'cmIdentityFactory', 'cmObject',
+    'cmStateManagement', 'cmUtil', 'cmLogger', 'cmSecurityAspectsContact', 'cmModal',
     '$q',
-    function(cmContactsAdapter, cmIdentityFactory, cmObject, cmStateManagement, cmUtil, cmLogger, cmSecurityAspectsContact, $q){
+    function(cmContactsAdapter, cmIdentityFactory, cmObject,
+             cmStateManagement, cmUtil, cmLogger, cmSecurityAspectsContact, cmModal,
+             $q){
         function ContactModel(data){
             var self = this;
 
@@ -110,12 +107,25 @@ angular.module('cmContacts')
                 return defer.promise;
             };
 
+            this.delete = function(withoutModal){
+                return withoutModal
+                    ? $q.when()
+                    : cmModal.confirm({
+                        title: 'CONTACT.MODAL.DELETE.HEADER',
+                        text: 'CONTACT.MODAL.DELETE.TEXT'
+                }).then(function() {
+                    return cmContactsAdapter
+                        .deleteContact(self.id)
+                }).then(function(){
+                    self.trigger('deleted:finished',self);
+                });
+            };
+
             init(data);
 
             this.identity.on('update:finished', function(){
                 self.securityAspects.scheduleRefresh();
             });
-
         }
 
         return ContactModel;