+++ core/comps/conversations/srvc-conversation-model.js
@@ -322,9 +322,13 @@ angular.module('cmConversations')
 
                             self.state.unset('loading');
                         },
-                        function(){
+                        function(response){
                             self.state.unset('loading');
                             self.trigger('load:failed');
+
+                            if(response.status && response.status == 404){
+                                self.trigger('notFound',self);
+                            }
                         }
                     );
                 } else if(this.state.is('loading') === false) {
@@ -440,6 +444,55 @@ angular.module('cmConversations')
                     this.update();
                 }
 
+            };
+
+            this.sendMessage = function(message, files, force){
+                if(!(message instanceof cmMessageModel)){
+                    return $q.reject('message no message-model');
+                }
+
+                return  this.getPassphrase()
+                        .catch(function(){
+                            return  self.isEncrypted()
+                                ?   $q.reject('access denied')
+                                :   $q.when(null);
+                            //Todo: null for 'not encrypted' old convention
+                        })
+                        .then(
+                            function(passphrase) {
+                                return self.isEncrypted()
+                                        ? message
+                                            .addFiles(files)
+                                            .getSignatures()
+                                            .then(function () {
+                                                return message.encrypt(passphrase)
+                                            })
+                                            .then(function () {
+                                                return message.save()
+                                            })
+
+                                        : message
+                                            .addFiles(files)
+                                            .setPublicData(['text', 'fileIds'])
+                                            .revealSignatures()
+                                            .getSignatures()
+                                            .then(function () {
+                                                return message.save()
+                                            })
+                            },
+                            function(){
+                                if(typeof force == 'boolean' && force === true){
+                                    return  message
+                                        .addFiles(files)
+                                        .setPublicData(['text', 'fileIds'])
+                                        .revealSignatures()
+                                        .getSignatures()
+                                        .then(function () {
+                                            return message.save()
+                                        })
+                                }
+                            }
+                        )
             };
 
             /**