+++ core/widgets/password/wdgt-password-code.js
@@ -67,7 +67,7 @@ angular.module('cmWidgets').directive('cmWidgetPasswordCode', [
                                 },
                                 function(response){
                                     loader.stop();
-                                    switch(response.data.errorCode){
+                                    switch(response.data.errorCodes[0]){
                                         case 'PASSWORD.RESET.EXPIRED':
                                             $scope.info.expired = true;
                                         break;