+++ test/lib/e2e/cmTestUtil.js
@@ -127,17 +127,29 @@ this.expectCurrentUrl = function (match) {
 this.logout = function () {
     self.get('/login')
 
-    return  $$("cm-menu").then(function (elements) {
-                if (elements.length > 0) {
-                    $("cm-menu .cm-handler").click()
-                    self.waitForElement(".cm-menu-list")
-                    .then(function(){
-                        return self.waitAndClickQa('logout-btn')
-                    });
-                }
-                return self.waitForPageLoad('/login')
-            })
-    
+    return $$("cm-menu")
+        .then(function (elements) {
+            if (elements.length > 0) {
+
+                $$("cm-search-input").then(function(elements){
+                    if(elements.length > 0){
+                        elements[0].getAttribute('class')
+                        .then(function(className){
+                            if(className.indexOf('visible') >= 0){
+                                self.click('btn-close-search')
+                            }
+                        })
+                    }
+                })
+
+                $("cm-menu .cm-handler").click()
+                self.waitForElement(".cm-menu-list")
+                .then(function(){
+                    return self.waitAndClickQa('logout-btn')
+                })
+            }
+            return self.waitForPageLoad('/login')
+        })
 }
 
 this.login = function (username, password, expectedRoute) {
@@ -176,24 +188,24 @@ this.createTestUser = function (testUserId, from){
     var password = 'password'
 
     this.get('/registration')
-    return  this.waitForPageLoad('/registration')
+    return this.waitForPageLoad('/registration')
             .then(function(){
                 self.setVal('input-cameoId',loginName,true)
                 self.setVal('input-password',password)
-                return  self.waitAndClickQa('icon-toggle-password')
-                        .then(function(){
-                            self.scrollToBottom()
-                            return self.waitAndClickQa('icon-checkbox-agb')
-                        })
-                        .then(function(){
-                            return self.waitAndClickQa('btn-createUser')                    
-                        })
-                        .then(function(){
-                            return self.waitForPageLoad("/setup/account")            
-                        })
-                        .then(function(){
-                            return loginName
-                        })
+                return self.waitAndClickQa('icon-toggle-password')
+                    .then(function(){
+                        self.scrollToBottom()
+                        return self.waitAndClickQa('icon-checkbox-agb')
+                    })
+                    .then(function(){
+                        return self.waitAndClickQa('btn-createUser')
+                    })
+                    .then(function(){
+                        return self.waitForPageLoad("/setup/account")
+                    })
+                    .then(function(){
+                        return loginName
+                    })
             })
 }
 
@@ -395,7 +407,7 @@ this.waitForEventSubscription = function () {
 }
 
 this.click = function (dataQa) {
-    $("[data-qa='" + dataQa + "']").click()
+    return $("[data-qa='" + dataQa + "']").click()
 }
 
 this.waitForQa = function(dataQa){
@@ -780,6 +792,48 @@ this.getVal = function (dataQa) {
 
 this.setValQuick = function (dataQa, text) {
     ptor.executeScript("document.querySelector(\"[data-qa='" + dataQa + "']\").value = '" + text + "'")
+}
+
+this.addRecipient = function(username){
+    self.waitAndClickQa('btn-add-recipients');
+    self.waitForPageLoad("/conversation/new/recipients")
+    self.headerSearchInList(username);
+    self.waitForQa('contact-display-name')
+    $("[data-qa='btn-select-contact']").click()
+
+    self.closeHeaderSearch()
+
+    self.get("/conversation/new")
+    self.waitForPageLoad("/conversation/new")
+}
+
+this.createUnencryptedConversation = function(subject, message, recpient){
+    self.get("/conversation/new")
+    self.waitForPageLoad("/conversation/new")
+
+    self.disableEncryption();
+
+    if(recpient){
+        if(typeof recpient == 'string'){
+            self.addRecipient(recpient)
+        }
+    }
+
+
+    self.setVal("input-subject", subject)
+    self.setVal("input-answer", message)
+    self.waitAndClickQa("btn-send-answer")
+    if(!recpient){
+        self.waitAndClickQa("btn-confirm", "cm-modal.active")
+    }
+    self.waitForPageLoad("/conversation/*")
+    self.waitForElements("cm-message", 1)
+
+    ptor.wait(function(){
+        return self.getVal("input-answer").then(function(answer){
+            return answer == ''
+        })
+    })
 }
 
 this.createEncryptedConversation = function (subject, message) {